+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix conversation cleanup when MessagingClientConversationJoinFailure occurs to prevent orphaned conversations
+
 ### Changed
 
 - Clean up `ChatSDK.sendTypingEvent()` unused code path